#3 - refactor GAV from org.apache.camel to com.github.cameltooling
