+++ .travis.yml
@@ -12,7 +12,7 @@ matrix:
         provider: releases
         api_key:  ${DEPLOY_TOKEN}
         file:
-          - "org.apache.camel.lsp.eclipse.updatesite/target/org.apache.camel.lsp.eclipse.updatesite-1.0.0-SNAPSHOT.zip"
+          - "com.github.camel-tooling.eclipse.updatesite/target/com.github.camel-tooling.eclipse.updatesite-1.0.0-SNAPSHOT.zip"
         skip_cleanup: true
         overwrite: true
         on:
