+++ Install.md
@@ -3,7 +3,7 @@ How to Install
 
 * Install Eclipse
 * Go to latest release page https://github.com/camel-idea-plugin/camel-lsp-client-eclipse/releases/latest/
-* Download zipped update site named org.apache.camel.lsp.eclipse.updatesite-1.0.0-SNAPSHOT.zip
+* Download zipped update site named com.github.camel-tooling.eclipse.updatesite-1.0.0-SNAPSHOT.zip
 * In Eclipse, Help -> Install New Software...
 * Click Add...
 * Click Archive...
