+++ camel-lsp-target-platform/pom.xml
@@ -5,7 +5,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	
 	<parent>
-		<groupId>org.apache.camel.tools.lsp</groupId>
+		<groupId>com.github.camel-tooling.lsp</groupId>
 		<artifactId>eclipse</artifactId>
 		<version>1.0.0-SNAPSHOT</version>
 	</parent>
