+++ com.github.camel-tooling.lsp.eclipse.client.tests.integration/META-INF/MANIFEST.MF
@@ -1,0 +1,16 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Bundle-SymbolicName: com.github.camel-tooling.eclipse.client.tests.integration;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Vendor: %Bundle-Vendor
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Bundle-ActivationPolicy: lazy
+Require-Bundle: org.junit;bundle-version="4.12.0",
+ org.assertj;bundle-version="1.7.1",
+ org.eclipse.lsp4e;bundle-version="0.3.0",
+ com.github.camel-tooling.eclipse.client;bundle-version="1.0.0",
+ org.eclipse.ui.ide;bundle-version="3.13.1",
+ org.eclipse.core.resources;bundle-version="3.12.0",
+ org.eclipse.core.runtime;bundle-version="3.13.0",
+ org.eclipse.ui.genericeditor;bundle-version="1.0.1"
