+++ com.github.camel-tooling.lsp.eclipse.client.tests.integration/infinitest.filters
