+++ com.github.camel-tooling.lsp.eclipse.client/.classpath
