+++ com.github.camel-tooling.lsp.eclipse.client/META-INF/MANIFEST.MF
@@ -1,0 +1,15 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: %Bundle-Name
+Bundle-SymbolicName: com.github.camel-tooling.eclipse.client;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Vendor: %Bundle-Vendor
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Bundle-ActivationPolicy: lazy
+Require-Bundle: org.eclipse.lsp4e;bundle-version="0.3.0",
+ org.eclipse.core.runtime;bundle-version="3.13.0",
+ org.eclipse.jface.text;bundle-version="3.12.0",
+ org.eclipse.wst.sse.ui;bundle-version="1.3.600"
+Export-Package: com.github.cameltooling.eclipse.client;x-friends:="com.github.camel-tooling.eclipse.client.tests.integration"
+Eclipse-BundleShape: dir
+Bundle-Activator: com.github.cameltooling.eclipse.client.ActivatorCamelLspClient
