+++ com.github.camel-tooling.lsp.eclipse.client/OSGI-INF/l10n/bundle.properties
