+++ com.github.camel-tooling.lsp.eclipse.client/build.properties
