+++ com.github.camel-tooling.lsp.eclipse.updatesite/.project
