+++ com.github.camel-tooling.lsp.eclipse.updatesite/category.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<site>
+   <bundle id="com.github.camel-tooling.eclipse.client" version="1.0.0.qualifier"/>
+</site>
