+++ com.github.camel-tooling.lsp.eclipse.updatesite/pom.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>com.github.camel-tooling.lsp</groupId>
+		<artifactId>eclipse</artifactId>
+		<version>1.0.0-SNAPSHOT</version>
+	</parent>
+	<artifactId>com.github.camel-tooling.eclipse.updatesite</artifactId>
+	<name>${base.name} :: Clients :: Eclipse Client :: Update Site</name>
+	<packaging>eclipse-repository</packaging>	
+</project>
