--- org.apache.camel.lsp.eclipse.client.tests.integration/pom.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.apache.camel.tools.lsp</groupId>
-		<artifactId>eclipse</artifactId>
-		<version>1.0.0-SNAPSHOT</version>
-	</parent>
-	<artifactId>org.apache.camel.lsp.eclipse.client.tests.integration</artifactId>
-	<name>${base.name} :: Clients :: Eclipse Client :: Integration Tests</name>
-	<packaging>eclipse-test-plugin</packaging>
-
-	<properties>
-		<tycho-version>0.26.0</tycho-version>
-	</properties>
-
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.eclipse.tycho</groupId>
-				<artifactId>tycho-surefire-plugin</artifactId>
-				<version>${tycho.version}</version>
-				<configuration>
-					<systemProperties combine.children="append">
-					</systemProperties>
-					<useUIThread>true</useUIThread>
-					<useUIHarness>true</useUIHarness>
-					<includes>
-						<include>**/*IT.class</include>
-					</includes>
-					<argLine>${tycho.testArgLine} -XX:+HeapDumpOnOutOfMemoryError</argLine>
-					<osgiDataDirectory>${project.basedir}/target/work sp@cé</osgiDataDirectory>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-</project>
