--- org.apache.camel.lsp.eclipse.client/META-INF/MANIFEST.MF
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: %Bundle-Name
-Bundle-SymbolicName: org.apache.camel.lsp.eclipse.client;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Bundle-Vendor: %Bundle-Vendor
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-ActivationPolicy: lazy
-Require-Bundle: org.eclipse.lsp4e;bundle-version="0.3.0",
- org.eclipse.core.runtime;bundle-version="3.13.0",
- org.eclipse.jface.text;bundle-version="3.12.0",
- org.eclipse.wst.sse.ui;bundle-version="1.3.600"
-Export-Package: org.apache.camel.lsp.eclipse.client;x-friends:="org.apache.camel.lsp.eclipse.client.tests.integration"
-Eclipse-BundleShape: dir
-Bundle-Activator: org.apache.camel.lsp.eclipse.client.ActivatorCamelLspClient
