--- org.apache.camel.lsp.eclipse.client/plugin.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.4"?>
-<plugin>
-   <extension
-         point="org.eclipse.lsp4e.languageServer">
-      <contentTypeMapping
-            contentType="org.eclipse.core.runtime.xml"
-            id="Camel LSP Server"
-            languageId="LANGUAGE_ID_APACHE_CAMEL">
-      </contentTypeMapping>
-      <server
-            class="org.apache.camel.lsp.eclipse.client.CamelLSPStreamConnectionProvider"
-            id="Camel LSP Server"
-            label="Camel LSP Server">
-      </server>
-   </extension>
-   <extension
-         point="org.eclipse.wst.sse.ui.completionProposal">
-         <proposalCategory
-           id="org.apache.camel.category"
-           name="Apache Camel completion proposals">
-  </proposalCategory>
- 
-  <proposalComputer
-       activate="true"
-       categoryId="org.apache.camel.category"
-       class="org.apache.camel.lsp.eclipse.xml.completion.CamelURICompletionProposalComputer"
-       id="org.apache.camel.xml.proposalcomputer">      
-    <contentType id="org.eclipse.core.runtime.xml"/>      
-  </proposalComputer> 
-   </extension>
-
-</plugin>
