--- org.apache.camel.lsp.eclipse.client/pom.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
-	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.apache.camel.tools.lsp</groupId>
-		<artifactId>eclipse</artifactId>
-		<version>1.0.0-SNAPSHOT</version>
-	</parent>
-	<artifactId>org.apache.camel.lsp.eclipse.client</artifactId>
-	<name>${base.name} :: Clients :: Eclipse Client :: Implementation</name>
-	<packaging>eclipse-plugin</packaging>
-	
-	<properties>
-		<tycho-version>0.26.0</tycho-version>
-	</properties>
-	
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-dependency-plugin</artifactId>
-				<executions>
-					<execution>
-						<id>get-deps</id>
-						<goals>
-							<goal>copy</goal>
-						</goals>
-						<phase>generate-resources</phase>
-					</execution>
-				</executions>
-				<configuration>
-					<artifactItems>
-						<artifactItem>
-							<groupId>org.apache.camel.tools.lsp</groupId>
-							<artifactId>camel-lsp-server</artifactId>
-							<version>${project.version}</version>
-						</artifactItem>
-					</artifactItems>
-					<useBaseVersion>true</useBaseVersion>
-					<skip>false</skip>
-					<outputDirectory>${basedir}/libs/</outputDirectory>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-</project>
