--- org.apache.camel.lsp.eclipse.updatesite/category.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<site>
-   <bundle id="org.apache.camel.lsp.eclipse.client" version="1.0.0.qualifier"/>
-</site>
