+++ pom.xml
@@ -4,7 +4,7 @@
 	xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>org.apache.camel.tools.lsp</groupId>
+	<groupId>com.github.camel-tooling.lsp</groupId>
 	<artifactId>eclipse</artifactId>
 	<version>1.0.0-SNAPSHOT</version>
 	<name>${base.name} :: Clients :: Eclipse Client</name>
@@ -29,9 +29,9 @@
 	
 	<modules>
 		<module>camel-lsp-target-platform</module>
-		<module>org.apache.camel.lsp.eclipse.client</module>
-		<module>org.apache.camel.lsp.eclipse.client.tests.integration</module>
-		<module>org.apache.camel.lsp.eclipse.updatesite</module>
+		<module>com.github.camel-tooling.lsp.eclipse.client</module>
+		<module>com.github.camel-tooling.lsp.eclipse.client.tests.integration</module>
+		<module>com.github.camel-tooling.lsp.eclipse.updatesite</module>
 	</modules>
 	
 	<build>
@@ -62,7 +62,7 @@
 				<version>0.7.9</version>
 				<configuration>
 					<append>true</append>
-					<includes>org.apache.camel.tools.lsp.*,org.apache.camel.lsp.*</includes>
+					<includes>com.github.cameltooling.lsp.*,com.github.cameltooling.*</includes>
 				</configuration>
 				<executions>
 					<execution>
